Move `host-config` to `config`
