+++ src/sidebar/config/fetch-config.js
@@ -1,5 +1,5 @@
 import getApiUrl from './get-api-url';
-import hostConfig from '../host-config';
+import hostConfig from './host-config';
 import * as postMessageJsonRpc from '../util/postmessage-json-rpc';
 
 /**
